+++ src/blocked.ts
@@ -18,8 +18,6 @@ const motivanionalMsgs = [
 let blockedUrl: string | undefined = '';
 
 document.addEventListener('DOMContentLoaded', () => {
-  console.log('DOMContentLoaded');
-  console.log(window.location.search);
   if (para) {
     getBlockedUrl();
   }
@@ -46,7 +44,6 @@ async function  getBlockedUrl() {
     const res: ResToSend = await browser.runtime.sendMessage(msg);
     if (res.success && res.rules) {
       blockedUrl = res.rules.find(rule => rule.id === urlId)?.strippedUrl;
-      console.log({blockedUrl});
       if (blockedUrl) {
         para!.textContent = blockedUrl;
         deleteBtn?.removeAttribute('disabled');
@@ -76,7 +73,6 @@ async function deleteRule(id: number) {
 }
 
 function displayMotivationMsg() {
-  console.log('displayMotivationMsg()');
   if (motivationHeading) {
     const idx = Math.floor(Math.random() * motivanionalMsgs.length);
     motivationHeading.innerText = motivanionalMsgs[idx];